agents/requirements.txt

google-adk
google-auth
google-genai>=1.9.0
google-cloud-aiplatform>=1.108.0
requests

llama-index>=0.10.0

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

pytest>=8.0.0
